requirements.txt
mypy
flake8
pytest

torch
torchvision
numpy

jupyterlab
matplotlib
tqdm
click
pyyaml
